+++ test/fixtures/win.js
@@ -242,7 +242,7 @@ export const escape = {
         expected: { interpolation: "a ", noInterpolation: "a " },
       },
       {
-        input: "a\n\rb",
+        input: "a\r\nb",
         expected: { interpolation: "a b", noInterpolation: "a b" },
       },
     ],
@@ -863,8 +863,8 @@ export const escape = {
         expected: { interpolation: "a", noInterpolation: "a" },
       },
       {
-        input: "a\n\rb",
-        expected: { interpolation: "a b", noInterpolation: "a\n\rb" },
+        input: "a\r\nb",
+        expected: { interpolation: "a b", noInterpolation: "a\r\nb" },
       },
     ],
     "<escape> ('\\u001B')": [