+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
+- Improve escaping performance when interpolation is set to `true`. ([#457])
 - Improve specificity of supported Node.js versions. ([#432])
 - Remove all side effects on import. ([#408])
 
@@ -182,5 +183,6 @@ Versioning].
 [#373]: https://github.com/ericcornelissen/shescape/pull/373
 [#408]: https://github.com/ericcornelissen/shescape/pull/408
 [#432]: https://github.com/ericcornelissen/shescape/pull/432
+[#457]: https://github.com/ericcornelissen/shescape/pull/457
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html