+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
+- Escape control characters `U+0008`, `U+000D`, `U+001B`, and `U+009B`. ([#456])
 - Improve escaping performance when interpolation is set to `true`. ([#457])
 - Improve specificity of supported Node.js versions. ([#432])
 - Remove all side effects on import. ([#408])
@@ -183,6 +184,7 @@ Versioning].
 [#373]: https://github.com/ericcornelissen/shescape/pull/373
 [#408]: https://github.com/ericcornelissen/shescape/pull/408
 [#432]: https://github.com/ericcornelissen/shescape/pull/432
+[#456]: https://github.com/ericcornelissen/shescape/pull/456
 [#457]: https://github.com/ericcornelissen/shescape/pull/457
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html