+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 
 - Escape control characters `U+0008`, `U+000D`, `U+001B`, and `U+009B`. ([#456])
 - Improve escaping performance when interpolation is set to `true`. ([#457])
+- Improve newline substitution on Windows. ([#458])
 - Improve specificity of supported Node.js versions. ([#432])
 - Remove all side effects on import. ([#408])
 
@@ -186,5 +187,6 @@ Versioning].
 [#432]: https://github.com/ericcornelissen/shescape/pull/432
 [#456]: https://github.com/ericcornelissen/shescape/pull/456
 [#457]: https://github.com/ericcornelissen/shescape/pull/457
+[#458]: https://github.com/ericcornelissen/shescape/pull/458
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html